+++ src/lib.rs
@@ -41,8 +41,9 @@
 //! `clock` and `std`; both are on by default.
 //!
 //! - The `clock` feature enables getting the current time using `Utc::now()` from [`chrono`].
-//!   Without it, some claim methods, such as [`Claims::set_duration()`], are not available.
-//!   It is still possible to set the corresponding claim fields manually.
+//!   Without it, some [`TimeOptions`] constructors, such as the `Default` impl,
+//!   are not available. It is still possible to create `TimeOptions` with an excplicitly specified
+//!   clock function, or to set / verify time-related [`Claims`] fields manually.
 //! - The `std` feature is propagated to the core dependencies and enables `std`-specific
 //!   functionality (such as error types implementing the standard `Error` trait).
 //!
@@ -86,15 +87,14 @@
 //! }
 //!
 //! # fn main() -> anyhow::Result<()> {
+//! // Choose time-related options for token creation / validation.
+//! let time_options = TimeOptions::default();
 //! // Create a symmetric HMAC key, which will be used both to create and verify tokens.
 //! let key = Hs256Key::from(b"super_secret_key_donut_steel" as &[_]);
 //! // Create a token.
-//! let header = Header {
-//!     key_id: Some("my-key".to_owned()),
-//!     ..Default::default()
-//! };
+//! let header = Header::default().with_key_id("my-key");
 //! let claims = Claims::new(CustomClaims { subject: "alice".to_owned() })
-//!     .set_duration_and_issuance(Duration::days(7))
+//!     .set_duration_and_issuance(&time_options, Duration::days(7))
 //!     .set_not_before(Utc::now() - Duration::hours(1));
 //! let token_string = Hs256.token(header, &claims, &key)?;
 //! println!("token: {}", token_string);
@@ -107,10 +107,9 @@
 //! // Validate the token integrity.
 //! let token: Token<CustomClaims> = Hs256.validate_integrity(&token, &key)?;
 //! // Validate additional conditions.
-//! token
-//!     .claims()
-//!     .validate_expiration(Leeway::default())?
-//!     .validate_maturity(Leeway::seconds(15))?;
+//! token.claims()
+//!     .validate_expiration(&time_options)?
+//!     .validate_maturity(&time_options)?;
 //! // Now, we can extract information from the token (e.g., its subject).
 //! let subject = &token.claims().custom.subject;
 //! assert_eq!(subject, "alice");
@@ -138,9 +137,10 @@
 //! }
 //!
 //! # fn main() -> anyhow::Result<()> {
+//! let time_options = TimeOptions::default();
 //! let key = Hs256Key::from(b"super_secret_key_donut_steel" as &[_]);
 //! let claims = Claims::new(CustomClaims { subject: [111; 32] })
-//!     .set_duration_and_issuance(Duration::days(7));
+//!     .set_duration_and_issuance(&time_options, Duration::days(7));
 //! let token = Hs256.token(Header::default(), &claims, &key)?;
 //! println!("token: {}", token);
 //! let compact_token = Hs256.compact_token(Header::default(), &claims, &key)?;
@@ -150,7 +150,7 @@
 //! // Parse the compact token.
 //! let token = UntrustedToken::try_from(compact_token.as_str())?;
 //! let token: Token<CustomClaims> = Hs256.validate_integrity(&token, &key)?;
-//! token.claims().validate_expiration(Leeway::default())?;
+//! token.claims().validate_expiration(&time_options)?;
 //! // Now, we can extract information from the token (e.g., its subject).
 //! assert_eq!(token.claims().custom.subject, [111; 32]);
 //! # Ok(())
@@ -199,13 +199,11 @@ mod alloc {
 
 /// Prelude to neatly import all necessary stuff from the crate.
 pub mod prelude {
-    pub use crate::{
-        AlgorithmExt as _, Claims, Header, Leeway, TimeOptions, Token, UntrustedToken,
-    };
+    pub use crate::{AlgorithmExt as _, Claims, Header, TimeOptions, Token, UntrustedToken};
 }
 
 pub use crate::{
-    claims::{Claims, Empty, Leeway, TimeOptions},
+    claims::{Claims, Empty, TimeOptions},
     error::{CreationError, ParseError, ValidationError},
 };
 
@@ -477,7 +475,18 @@ impl<A: Algorithm> AlgorithmExt for A {
 /// the verifying key. Since these values will be provided by the adversary in the case of
 /// an attack, they require additional verification (e.g., a provided certificate might
 /// be checked against the list of "acceptable" certificate authorities).
+///
+/// A `Header` can be created using `Default` implementation, which does not set any fields.
+/// For added fluency, you may use `with_*` methods:
+///
+/// ```
+/// # use jwt_compact::Header;
+/// let header = Header::default()
+///     .with_key_id("my-key-id")
+///     .with_certificate_thumbprint("thumbprint");
+/// ```
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
+#[non_exhaustive]
 pub struct Header {
     /// URL of the JSON Web Key Set containing the key that has signed the token.
     /// This field is renamed to `jku` for serialization.
@@ -502,6 +511,41 @@ pub struct Header {
     /// Application-specific signature type. This field is renamed to `typ` for serialization.
     #[serde(rename = "typ", default, skip_serializing_if = "Option::is_none")]
     pub signature_type: Option<String>,
+}
+
+impl Header {
+    /// Sets the `key_set_url` field for this instance.
+    pub fn with_key_set_url(mut self, key_set_url: impl Into<String>) -> Self {
+        self.key_set_url = Some(key_set_url.into());
+        self
+    }
+
+    /// Sets the `key_id` field for this instance.
+    pub fn with_key_id(mut self, key_id: impl Into<String>) -> Self {
+        self.key_id = Some(key_id.into());
+        self
+    }
+
+    /// Sets the `certificate_url` field for this instance.
+    pub fn with_certificate_url(mut self, certificate_url: impl Into<String>) -> Self {
+        self.certificate_url = Some(certificate_url.into());
+        self
+    }
+
+    /// Sets the `certificate_thumbprint` field for this instance.
+    pub fn with_certificate_thumbprint(
+        mut self,
+        certificate_thumbprint: impl Into<String>,
+    ) -> Self {
+        self.certificate_thumbprint = Some(certificate_thumbprint.into());
+        self
+    }
+
+    /// Sets the `signature_type` field for this instance.
+    pub fn with_signature_type(mut self, signature_type: impl Into<String>) -> Self {
+        self.signature_type = Some(signature_type.into());
+        self
+    }
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -573,7 +617,8 @@ impl<T> Token<T> {
 ///
 /// # fn main() -> anyhow::Result<()> {
 /// # let key = Hs256Key::from(b"super_secret_key" as &[_]);
-/// # let claims = Claims::new(MyClaims {}).set_duration_and_issuance(Duration::days(7));
+/// # let claims = Claims::new(MyClaims {})
+/// #     .set_duration_and_issuance(&TimeOptions::default(), Duration::days(7));
 /// let token_string: String = // token from an external source
 /// #   Hs256.token(Header::default(), &claims, &key)?;
 /// let token = UntrustedToken::try_from(token_string.as_str())?;
@@ -583,7 +628,7 @@ impl<T> Token<T> {
 /// let array: GenericArray<u8, typenum::U32> = signed.signature.into_bytes();
 /// // Token itself is available via `token` field.
 /// let claims = signed.token.claims();
-/// claims.validate_expiration(Leeway::default())?;
+/// claims.validate_expiration(&TimeOptions::default())?;
 /// // Process the claims...
 /// # Ok(())
 /// # } // end main()